Make PartitionPublishingOptions and PartitionPublishingProperties private
