+++ sdk/eventhub/event-hubs/src/index.ts
@@ -16,8 +16,6 @@ export {
   GetPartitionIdsOptions,
   GetPartitionPropertiesOptions,
   GetEventHubPropertiesOptions,
-  PartitionPublishingOptions,
-  PartitionPublishingProperties,
 } from "./models/public";
 export { EventHubConsumerClient } from "./eventHubConsumerClient";
 export { EventHubProducerClient } from "./eventHubProducerClient";
