+++ sdk/eventhub/event-hubs/src/models/public.ts
@@ -179,44 +179,6 @@ export interface EventHubConsumerClientOptions extends EventHubClientOptions {
 }
 
 /**
- * The set of options that can be specified for an `EventHubProducerClient`
- * to influence its behavior when publishing directly to an Event Hub partition.
- *
- * These options are ignored when publishing to the Event Hubs gateway for automatic
- * routing or when using a partition key.
- */
-export interface PartitionPublishingOptions {
-  /**
-   * The owner level indicates that a publishing is intended to be performed exclusively for events in the
-   * requested partition in the context of the associated producer group.
-   *
-   * To do so, publishing will attempt to assert ownership over the partition;
-   * in the case where more than one publisher in the producer group attempts to assert ownership for the same partition,
-   * the one having a larger `ownerLevel` value will "win".
-   */
-  ownerLevel?: number;
-  /**
-   * The identifier of the producer group that this producer is associated with when publishing to the associated partition.
-   * Events will be published in the context of this group.
-   *
-   * If `undefined`, the Event Hubs service will control the value.
-   *
-   * The producer group is only recognized and relevant when certain features of the producer are enabled.
-   * For example, it is used by idempotent publishing.
-   */
-  producerGroupId?: number;
-  /**
-   * The starting number that should be used for the automatic sequencing of events for the associated partition, when published by this producer.
-   *
-   * If `undefined`, the Event Hubs service will control the value.
-   *
-   * The producer group is only recognized and relevant when certain features of the producer are enabled.
-   * For example, it is used by idempotent publishing.
-   */
-  startingSequenceNumber?: number;
-}
-
-/**
  * An options bag to configure load balancing settings.
  */
 export interface LoadBalancingOptions {
@@ -280,30 +242,4 @@ export interface CreateBatchOptions extends OperationOptions {
    * The upper limit for the size of batch. The `tryAdd` function will return `false` after this limit is reached.
    */
   maxSizeInBytes?: number;
-}
-
-/**
- * Describes the information about the state of publishing for a partition.
- */
-export interface PartitionPublishingProperties {
-  /**
-   * Indicates whether or not idempotent publishing is enabled for the producer and, by extension, the associated partition.
-   */
-  isIdempotentPublishingEnabled: boolean;
-  /**
-   * The owner level of the producer publishing to the associated partition.
-   */
-  ownerLevel?: number;
-  /**
-   * The partition id the properties are associated with.
-   */
-  partitionId: string;
-  /**
-   * The identifier of the producer group for which this producer is publishing to the associated partition.
-   */
-  producerGroupId?: number;
-  /**
-   * The sequence number assigned to the event that was most recently published to the associated partition successfully.
-   */
-  lastPublishedSequenceNumber?: number;
 }
